code cleaning

+++ src/main/java/com/wtchuigo/reunion/services/MemberServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.wtchuigo.reunion.core.MemberDto;
+import com.wtchuigo.reunion.exceptions.ResourceNotFoundException;
 import com.wtchuigo.reunion.mapper.MemberMapper;
 import com.wtchuigo.reunion.model.Member;
 import com.wtchuigo.reunion.repositories.MemberRepository;
@@ -32,7 +33,10 @@ public class MemberServiceImpl implements MemberService {
 
 	@Override
 	public void delete(int id) {
-		memberRepository.deleteById(id);;
+		if (!memberRepository.existsById(id)) {
+			throw new ResourceNotFoundException("No member found with the id: " + id);
+		}
+		memberRepository.deleteById(id);
 	}
 
 	@Override
